Move "special" logging util to rest
