+++ twister-lib-android/logging/src/main/java/net/twisterrob/android/utils/log/LoggingCursorFactory.java
@@ -1,4 +1,4 @@
-package net.twisterrob.android.db;
+package net.twisterrob.android.utils.log;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
